remove logic
