+++ .gitignore
@@ -3,3 +3,4 @@
 etc
 .rspec_status
 tmp
+.zed
